backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.3
pydantic==2.9.0
pygithub==2.3.0
gitpython==3.1.43
httpx==0.27.2
openai==1.47.0
python-dotenv==1.0.1
semgrep==1.77.0
